+++ src/js/_init.js
@@ -13,8 +13,11 @@
 			EqUI.side_nav.update();
 
 			EqUI.buttons.update();
+			EqUI.cards.update();
 			EqUI.forms.update();
 			EqUI.collapsible.update();
+			EqUI.modals.update();
+			EqUI.tabs.update();
 			
 			EqUI.site.update();
 
@@ -27,6 +30,9 @@
 
 	        // Show body
             EqUI.site.body.css('visibility', 'visible');
+
+            // Loads
+            EqUI.tabs.load();
 	    });
 	    
 	});