+++ src/js/layout/side-nav.js
@@ -1,14 +1,21 @@
 (function ($) {
     EqUI.side_nav = {};
 
+    EqUI.side_nav.element = $('.eq-ui-side-nav');
+    EqUI.side_nav.header_element = $('.eq-ui-side-nav .brand');
+    EqUI.side_nav.footer_element = $('.eq-ui-side-nav .footer');
+    EqUI.side_nav.content_wrapper_element = $('.eq-ui-side-nav .eq-ui-side-nav-content-wrapper');
+
     // Init
     EqUI.side_nav.init = function() {
-        // Global vars
-        EqUI.side_nav.element = $('.eq-ui-side-nav');
 
         $('.eq-ui-side-nav-toggle').on('click', function(){
             EqUI.side_nav.nav_toggle();
         });
+
+        // Calculate height for (eq-ui-side-nav-content-wrapper)
+        var _content_wrapper_height_offset = EqUI.side_nav.header_element.outerHeight() + EqUI.side_nav.footer_element.outerHeight();
+        EqUI.side_nav.content_wrapper_element.css('height','calc(100% - '+_content_wrapper_height_offset+'px)');
     };
 
     // Update