+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 ## CHANGELOG
 
+### v0.2.1 (Nov ?, 2015)
+
+**New**
+- Create [dropdown](http://bootstrap-material-ui.exentriq.com/dropdown.html)
+- Create [tabs](http://bootstrap-material-ui.exentriq.com/tabs.html)
+
+### v0.2.0 (Nov 17, 2015)
+
+**New**
+- Create [cards](http://bootstrap-material-ui.exentriq.com/cards.html)
+- Create [list](http://bootstrap-material-ui.exentriq.com/list.html)
+- Create [modals](http://bootstrap-material-ui.exentriq.com/modals.html)
+- Create button fab floating
+- Create Group level for side bar menu
+- Create 3º level for side bar menu
+- Create top search component
+- Create search component
+- Create top nav menu component
+
+**Updates**
+- Fixed Header and Footer in Side Bar
+- Re-Work Collapside
+- Add new icons font, [Material Design Icons](https://materialdesignicons.com/)
+- New input classes
+- Selected all checkboxes in table
+
+**Fixes**
+- mangle to flase in uglify
+
 ### v0.1.4 (Sept 11, 2015)
 
 **New**